# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Decimal": {
				"type": "string"
			},
			"_36_Enums.TriagePriority": {
				"type": "string",
				"enum": [
					"RED",
					"YELLOW",
					"GREEN",
					"BLACK"
				]
			},
...
			"_36_Enums.VisitType": {
				"type": "string",
				"enum": [
					"OPD",
					"EMERGENCY",
					"REFERRAL",
					"FOLLOW_UP"
				]
			},
...
			"_36_Enums.VisitPriority": {
				"type": "string",
				"enum": [
					"NORMAL",
					"URGENT",
					"CRITICAL"
				]
			},
...
			"CreateMedicalServiceInput": {
				"properties": {
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"category",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreateMedicalServiceInput_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateMedicalServiceInput": {
				"$ref": "#/components/schemas/Partial_CreateMedicalServiceInput_"
			},
			"Pick_ProvideServiceInput.Exclude_keyofProvideServiceInput.providedById__": {
				"properties": {
					"visitId": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					},
					"serviceId": {
						"type": "string"
					}
				},
				"required": [
					"visitId",
					"serviceId"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_ProvideServiceInput.providedById_": {
				"$ref": "#/components/schemas/Pick_ProvideServiceInput.Exclude_keyofProvideServiceInput.providedById__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "auth-service",
		"version": "1.0.0",
		"description": "Authentication and RBAC Microservice",
		"contact": {}
	},
	"paths": {
...
		"/api/v1/medical-services": {
			"post": {
				"operationId": "CreateService",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"price": {
											"$ref": "#/components/schemas/Decimal"
										},
										"category": {
											"type": "string"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"id": {
											"type": "string"
										},
										"name": {
											"type": "string"
										}
									},
									"required": [
										"price",
										"category",
										"updatedAt",
										"createdAt",
										"id",
										"name"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"409": {
						"description": "Service Already Exists"
					}
				},
				"description": "Create a new medical service definition (Admin only)",
				"tags": [
					"Medical Services"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMedicalServiceInput"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetServices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"meta": {
											"properties": {
												"totalPages": {
													"type": "number",
													"format": "double"
												},
												"limit": {
													"type": "number",
													"format": "double"
												},
												"page": {
													"type": "number",
													"format": "double"
												},
												"total": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"totalPages",
												"limit",
												"page",
												"total"
											],
											"type": "object"
										},
										"data": {
											"items": {
												"properties": {
													"price": {
														"$ref": "#/components/schemas/Decimal"
													},
													"category": {
														"type": "string"
													},
													"updatedAt": {
														"type": "string",
														"format": "date-time"
													},
													"createdAt": {
														"type": "string",
														"format": "date-time"
													},
													"id": {
														"type": "string"
													},
													"name": {
														"type": "string"
													}
												},
												"required": [
													"price",
													"category",
													"updatedAt",
													"createdAt",
													"id",
													"name"
												],
												"type": "object"
											},
											"type": "array"
										}
									},
									"required": [
										"meta",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Retrieve a paginated and searchable list of medical services",
				"tags": [
					"Medical Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/medical-services/{id}": {
			"get": {
				"operationId": "GetServiceById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"price": {
											"$ref": "#/components/schemas/Decimal"
										},
										"category": {
											"type": "string"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"id": {
											"type": "string"
										},
										"name": {
											"type": "string"
										}
									},
									"required": [
										"price",
										"category",
										"updatedAt",
										"createdAt",
										"id",
										"name"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Medical Service Not Found"
					}
				},
				"description": "Retrieve a single medical service by ID",
				"tags": [
					"Medical Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"price": {
											"$ref": "#/components/schemas/Decimal"
										},
										"category": {
											"type": "string"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"id": {
											"type": "string"
										},
										"name": {
											"type": "string"
										}
									},
									"required": [
										"price",
										"category",
										"updatedAt",
										"createdAt",
										"id",
										"name"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Medical Service Not Found"
					}
				},
				"description": "Update medical service details (Admin only)",
				"tags": [
					"Medical Services"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMedicalServiceInput"
							}
						}
					}
				}
			}
		},
		"/api/v1/medical-services/provide": {
			"post": {
				"operationId": "ProvideService",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"unitPrice": {
											"$ref": "#/components/schemas/Decimal"
										},
										"providedById": {
											"type": "string"
										},
										"serviceId": {
											"type": "string"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"notes": {
											"type": "string"
										},
										"visitId": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"unitPrice",
										"providedById",
										"serviceId",
										"updatedAt",
										"createdAt",
										"notes",
										"visitId",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Patient Visit or Medical Service Not Found"
					}
				},
				"description": "Record a service provided to a patient visit",
				"tags": [
					"Medical Services"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_ProvideServiceInput.providedById_"
							}
						}
					}
				}
			}
		},
		"/api/v1/medical-services/patient/mrn/{mrn}": {
			"get": {
				"operationId": "GetByMrn",
				"responses": {
					"200": {
						"description": "Provided services retrieved by MRN",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"allOf": [
											{
												"properties": {
													"service": {
														"properties": {
															"price": {
																"$ref": "#/components/schemas/Decimal"
															},
															"category": {
																"type": "string"
															},
															"id": {
																"type": "string"
															},
															"name": {
																"type": "string"
															}
														},
														"required": [
															"price",
															"category",
															"id",
															"name"
														],
														"type": "object"
													},
													"visit": {
														"properties": {
															"createdAt": {
																"type": "string",
																"format": "date-time"
															},
															"id": {
																"type": "string"
															}
														},
														"required": [
															"createdAt",
															"id"
														],
														"type": "object"
													}
												},
												"required": [
													"service",
													"visit"
												],
												"type": "object"
											},
											{
												"properties": {
													"unitPrice": {
														"$ref": "#/components/schemas/Decimal"
													},
													"providedById": {
														"type": "string"
													},
													"serviceId": {
														"type": "string"
													},
													"updatedAt": {
														"type": "string",
														"format": "date-time"
													},
													"createdAt": {
														"type": "string",
														"format": "date-time"
													},
													"notes": {
														"type": "string"
													},
													"visitId": {
														"type": "string"
													},
													"id": {
														"type": "string"
													}
												},
												"required": [
													"unitPrice",
													"providedById",
													"serviceId",
													"updatedAt",
													"createdAt",
													"notes",
													"visitId",
													"id"
												],
												"type": "object"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all provided services for a patient using their MRN",
				"tags": [
					"Medical Services"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "mrn",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/medical-services/visit/latest/mrn/{mrn}": {
			"get": {
				"operationId": "GetLatestVisitByMrn",
				"responses": {
					"200": {
						"description": "Latest visit retrieved by MRN",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"properties": {
												"services": {
													"items": {
														"allOf": [
															{
																"properties": {
																	"service": {
																		"properties": {
																			"price": {
																				"$ref": "#/components/schemas/Decimal"
																			},
																			"category": {
																				"type": "string"
																			},
																			"updatedAt": {
																				"type": "string",
																				"format": "date-time"
																			},
																			"createdAt": {
																				"type": "string",
																				"format": "date-time"
																			},
																			"id": {
																				"type": "string"
																			},
																			"name": {
																				"type": "string"
																			}
																		},
																		"required": [
																			"price",
																			"category",
																			"updatedAt",
																			"createdAt",
																			"id",
																			"name"
																		],
																		"type": "object"
																	}
																},
																"required": [
																	"service"
																],
																"type": "object"
															},
															{
																"properties": {
																	"unitPrice": {
																		"$ref": "#/components/schemas/Decimal"
																	},
																	"providedById": {
																		"type": "string"
																	},
																	"serviceId": {
																		"type": "string"
																	},
																	"updatedAt": {
																		"type": "string",
																		"format": "date-time"
																	},
																	"createdAt": {
																		"type": "string",
																		"format": "date-time"
																	},
																	"notes": {
																		"type": "string"
																	},
																	"visitId": {
																		"type": "string"
																	},
																	"id": {
																		"type": "string"
																	}
																},
																"required": [
																	"unitPrice",
																	"providedById",
																	"serviceId",
																	"updatedAt",
																	"createdAt",
																	"notes",
																	"visitId",
																	"id"
																],
																"type": "object"
															}
														]
													},
													"type": "array"
												}
											},
											"required": [
												"services"
											],
											"type": "object"
										},
										{
											"properties": {
												"triagePriority": {
													"$ref": "#/components/schemas/_36_Enums.TriagePriority"
												},
												"visitType": {
													"$ref": "#/components/schemas/_36_Enums.VisitType"
												},
												"symptoms": {
													"type": "string"
												},
												"icdCode": {
													"type": "string"
												},
												"diagnosis": {
													"type": "string"
												},
												"attendingId": {
													"type": "string"
												},
												"patientId": {
													"type": "string"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"priority": {
													"$ref": "#/components/schemas/_36_Enums.VisitPriority"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"triagePriority",
												"visitType",
												"symptoms",
												"icdCode",
												"diagnosis",
												"attendingId",
												"patientId",
												"createdAt",
												"priority",
												"id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "Get the latest visit and its provided services by MRN",
				"tags": [
					"Medical Services"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "mrn",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}